update helpdesk system

+++ internal/repositories/position-repository.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"net/http"
 
-
 	sq "github.com/Masterminds/squirrel"
 
 	"github.com/jackc/pgx/v5"
@@ -17,8 +16,8 @@ import (
 
 	"request-system/internal/entities"
 	// ВАЖНО: Подключаем билдер
-	"request-system/internal/infrastructure/bd" 
-	
+	"request-system/internal/infrastructure/bd"
+
 	apperrors "request-system/pkg/errors"
 	"request-system/pkg/types"
 )
@@ -46,7 +45,7 @@ type PositionRepositoryInterface interface {
 	GetAll(ctx context.Context, filter types.Filter) ([]*entities.Position, uint64, error)
 	Create(ctx context.Context, tx pgx.Tx, p entities.Position) (uint64, error)
 	Update(ctx context.Context, tx pgx.Tx, id uint64, p entities.Position) error
-	Delete(ctx context.Context, tx pgx.Tx, id int) error
+	Delete(ctx context.Context, tx pgx.Tx, id uint64) error
 
 	FindByExternalID(ctx context.Context, tx pgx.Tx, externalID string, sourceSystem string) (*entities.Position, error)
 	FindByName(ctx context.Context, tx pgx.Tx, name string) (*entities.Position, error)
@@ -63,11 +62,10 @@ func NewPositionRepository(storage *pgxpool.Pool, logger *zap.Logger) PositionRe
 	return &positionRepository{storage: storage, logger: logger}
 }
 
-// -------------------------------------------------------------
-// Хелперы (Querier, Scan, FindOne)
-// -------------------------------------------------------------
 func (r *positionRepository) getQuerier(tx pgx.Tx) Querier {
-	if tx != nil { return tx }
+	if tx != nil {
+		return tx
+	}
 	return r.storage
 }
 
@@ -81,11 +79,17 @@ func (r *positionRepository) scanRow(row pgx.Row) (*entities.Position, error) {
 		&externalID, &sourceSystem,
 	)
 	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) { return nil, apperrors.ErrNotFound }
+		if errors.Is(err, pgx.ErrNoRows) {
+			return nil, apperrors.ErrNotFound
+		}
 		return nil, fmt.Errorf("ошибка сканирования positions: %w", err)
 	}
-	if externalID.Valid { p.ExternalID = &externalID.String }
-	if sourceSystem.Valid { p.SourceSystem = &sourceSystem.String }
+	if externalID.Valid {
+		p.ExternalID = &externalID.String
+	}
+	if sourceSystem.Valid {
+		p.SourceSystem = &sourceSystem.String
+	}
 	return &p, nil
 }
 
@@ -98,12 +102,9 @@ func (r *positionRepository) findOnePosition(ctx context.Context, querier Querie
 	return r.scanRow(querier.QueryRow(ctx, query, args...))
 }
 
-// -------------------------------------------------------------
-// GetAll - ЧЕРЕЗ BD HELPER
-// -------------------------------------------------------------
 func (r *positionRepository) GetAll(ctx context.Context, filter types.Filter) ([]*entities.Position, uint64, error) {
 	psql := sq.StatementBuilder.PlaceholderFormat(sq.Dollar)
-	
+
 	// Ручной поиск (OR ILIKE)
 	applySearch := func(b sq.SelectBuilder) sq.SelectBuilder {
 		if filter.Search != "" {
@@ -124,13 +125,17 @@ func (r *positionRepository) GetAll(ctx context.Context, filter types.Filter) ([
 	countBuilder = bd.ApplyListParams(countBuilder, countFilter, positionMap)
 
 	countQuery, countArgs, err := countBuilder.ToSql()
-	if err != nil { return nil, 0, fmt.Errorf("ошибка сборки SQL count: %w", err) }
+	if err != nil {
+		return nil, 0, fmt.Errorf("ошибка сборки SQL count: %w", err)
+	}
 
 	var total uint64
 	if err := r.storage.QueryRow(ctx, countQuery, countArgs...).Scan(&total); err != nil {
 		return nil, 0, fmt.Errorf("ошибка выполнения count: %w", err)
 	}
-	if total == 0 { return []*entities.Position{}, 0, nil }
+	if total == 0 {
+		return []*entities.Position{}, 0, nil
+	}
 
 	// 2. SELECT
 	selectBuilder := psql.Select(positionFields).From(positionTable)
@@ -144,27 +149,32 @@ func (r *positionRepository) GetAll(ctx context.Context, filter types.Filter) ([
 	selectBuilder = bd.ApplyListParams(selectBuilder, filter, positionMap)
 
 	query, args, err := selectBuilder.ToSql()
-	if err != nil { return nil, 0, fmt.Errorf("ошибка сборки SQL select: %w", err) }
+	if err != nil {
+		return nil, 0, fmt.Errorf("ошибка сборки SQL select: %w", err)
+	}
 
 	rows, err := r.storage.Query(ctx, query, args...)
-	if err != nil { return nil, 0, fmt.Errorf("ошибка выполнения select: %w", err) }
+	if err != nil {
+		return nil, 0, fmt.Errorf("ошибка выполнения select: %w", err)
+	}
 	defer rows.Close()
 
 	positions := make([]*entities.Position, 0)
 	for rows.Next() {
 		pos, err := r.scanRow(rows)
-		if err != nil { return nil, 0, err }
+		if err != nil {
+			return nil, 0, err
+		}
 		positions = append(positions, pos)
 	}
 
-	if err := rows.Err(); err != nil { return nil, 0, fmt.Errorf("ошибка итерации rows: %w", err) }
+	if err := rows.Err(); err != nil {
+		return nil, 0, fmt.Errorf("ошибка итерации rows: %w", err)
+	}
 
 	return positions, total, nil
 }
 
-// -------------------------------------------------------------
-// CRUD
-// -------------------------------------------------------------
 func (r *positionRepository) FindByID(ctx context.Context, tx pgx.Tx, id uint64) (*entities.Position, error) {
 	return r.findOnePosition(ctx, r.getQuerier(tx), sq.Eq{"id": id})
 }
@@ -177,11 +187,15 @@ func (r *positionRepository) FindByExternalID(ctx context.Context, tx pgx.Tx, ex
 func (r *positionRepository) FindIDByType(ctx context.Context, tx pgx.Tx, typeName string) (uint64, error) {
 	psql := sq.StatementBuilder.PlaceholderFormat(sq.Dollar)
 	query, args, err := psql.Select("id").From(positionTable).Where(sq.Eq{"type": typeName}).Limit(1).ToSql()
-	if err != nil { return 0, fmt.Errorf("ошибка сборки FindIDByType: %w", err) }
+	if err != nil {
+		return 0, fmt.Errorf("ошибка сборки FindIDByType: %w", err)
+	}
 
 	var id uint64
 	if err := r.getQuerier(tx).QueryRow(ctx, query, args...).Scan(&id); err != nil {
-		if errors.Is(err, pgx.ErrNoRows) { return 0, apperrors.ErrNotFound }
+		if errors.Is(err, pgx.ErrNoRows) {
+			return 0, apperrors.ErrNotFound
+		}
 		return 0, err
 	}
 	return id, nil
@@ -193,7 +207,9 @@ func (r *positionRepository) Create(ctx context.Context, tx pgx.Tx, p entities.P
 		Columns("name", "department_id", "otdel_id", "branch_id", "office_id", "type", "status_id", "external_id", "source_system", "created_at", "updated_at").
 		Values(p.Name, p.DepartmentID, p.OtdelID, p.BranchID, p.OfficeID, p.Type, p.StatusID, p.ExternalID, p.SourceSystem, sq.Expr("NOW()"), sq.Expr("NOW()")).
 		Suffix("RETURNING id").ToSql()
-	if err != nil { return 0, fmt.Errorf("ошибка сборки Create: %w", err) }
+	if err != nil {
+		return 0, fmt.Errorf("ошибка сборки Create: %w", err)
+	}
 
 	var newID uint64
 	if err := tx.QueryRow(ctx, query, args...).Scan(&newID); err != nil {
@@ -218,7 +234,9 @@ func (r *positionRepository) Update(ctx context.Context, tx pgx.Tx, id uint64, p
 		Set("status_id", p.StatusID).
 		Set("updated_at", sq.Expr("NOW()")).
 		Where(sq.Eq{"id": id}).ToSql()
-	if err != nil { return fmt.Errorf("ошибка сборки Update: %w", err) }
+	if err != nil {
+		return fmt.Errorf("ошибка сборки Update: %w", err)
+	}
 
 	result, err := tx.Exec(ctx, query, args...)
 	if err != nil {
@@ -228,14 +246,18 @@ func (r *positionRepository) Update(ctx context.Context, tx pgx.Tx, id uint64, p
 		}
 		return fmt.Errorf("ошибка обновления positions: %w", err)
 	}
-	if result.RowsAffected() == 0 { return apperrors.ErrNotFound }
+	if result.RowsAffected() == 0 {
+		return apperrors.ErrNotFound
+	}
 	return nil
 }
 
-func (r *positionRepository) Delete(ctx context.Context, tx pgx.Tx, id int) error {
+func (r *positionRepository) Delete(ctx context.Context, tx pgx.Tx, id uint64) error {
 	psql := sq.StatementBuilder.PlaceholderFormat(sq.Dollar)
 	query, args, err := psql.Delete(positionTable).Where(sq.Eq{"id": id}).ToSql()
-	if err != nil { return fmt.Errorf("ошибка сборки Delete: %w", err) }
+	if err != nil {
+		return fmt.Errorf("ошибка сборки Delete: %w", err)
+	}
 
 	result, err := tx.Exec(ctx, query, args...)
 	if err != nil {
@@ -245,7 +267,9 @@ func (r *positionRepository) Delete(ctx context.Context, tx pgx.Tx, id int) erro
 		}
 		return fmt.Errorf("ошибка удаления: %w", err)
 	}
-	if result.RowsAffected() == 0 { return apperrors.ErrNotFound }
+	if result.RowsAffected() == 0 {
+		return apperrors.ErrNotFound
+	}
 	return nil
 }
 
@@ -253,20 +277,30 @@ func (r *positionRepository) FindByTypeAndOrg(ctx context.Context, tx pgx.Tx, po
 	psql := sq.StatementBuilder.PlaceholderFormat(sq.Dollar)
 	builder := psql.Select(positionFields).From(positionTable).Where(sq.Eq{"type": posType}).OrderBy("id")
 
-	if depID != nil { builder = builder.Where(sq.Eq{"department_id": *depID}) }
-	if otdelID != nil { builder = builder.Where(sq.Eq{"otdel_id": *otdelID}) }
+	if depID != nil {
+		builder = builder.Where(sq.Eq{"department_id": *depID})
+	}
+	if otdelID != nil {
+		builder = builder.Where(sq.Eq{"otdel_id": *otdelID})
+	}
 
 	query, args, err := builder.ToSql()
-	if err != nil { return nil, fmt.Errorf("ошибка сборки FindByTypeAndOrg: %w", err) }
+	if err != nil {
+		return nil, fmt.Errorf("ошибка сборки FindByTypeAndOrg: %w", err)
+	}
 
 	rows, err := r.getQuerier(tx).Query(ctx, query, args...)
-	if err != nil { return nil, fmt.Errorf("ошибка выполнения: %w", err) }
+	if err != nil {
+		return nil, fmt.Errorf("ошибка выполнения: %w", err)
+	}
 	defer rows.Close()
 
 	var positions []*entities.Position
 	for rows.Next() {
 		pos, err := r.scanRow(rows)
-		if err != nil { return nil, err }
+		if err != nil {
+			return nil, err
+		}
 		positions = append(positions, pos)
 	}
 	return positions, nil
